@@ -4,6 +4,10 @@ version: 3.0.0
 author: Iiro Krankka <iiro.krankka@gmail.com>
 homepage: https://github.com/roughike/flutter_facebook_login
 
+
+environment:
+  sdk: ">=2.12.0-259.8.beta <3.0.0"
+  
 dependencies:
   flutter:
     sdk: flutter
@@ -18,5 +22,4 @@ flutter:
     androidPackage: com.roughike.facebooklogin.facebooklogin
     pluginClass: FacebookLoginPlugin
 
-environment:
-  sdk: ">=2.0.0-dev.28.0 <3.0.0"
+
