@@ -82,8 +82,8 @@ class FacebookLogin {
   ///
   /// NOTE: This might return an access token that has expired. If you need to be
   /// sure that the token is still valid, call [isValid] on the access token.
-  Future<FacebookAccessToken> get currentAccessToken async {
-    final Map<dynamic, dynamic> accessToken =
+  Future<FacebookAccessToken?> get currentAccessToken async {
+    final Map<dynamic, dynamic>? accessToken =
         await channel.invokeMethod('getCurrentAccessToken');
 
     if (accessToken == null) {
@@ -105,10 +105,10 @@ class FacebookLogin {
     List<String> permissions,
   ) async {
     final Map<dynamic, dynamic> result =
-        await channel.invokeMethod('logIn', {
+        await (channel.invokeMethod('logIn', {
       'behavior': _currentLoginBehaviorAsString(),
       'permissions': permissions,
-    });
+    }) as FutureOr<Map<dynamic, dynamic>>);
 
     return _deliverResult(
         FacebookLoginResult._(result.cast<String, dynamic>()));
@@ -212,13 +212,13 @@ class FacebookLoginResult {
   ///
   /// Only available when the [status] equals [FacebookLoginStatus.loggedIn],
   /// otherwise null.
-  final FacebookAccessToken accessToken;
+  final FacebookAccessToken? accessToken;
 
   /// The error message when the log in flow completed with an error.
   ///
   /// Only available when the [status] equals [FacebookLoginStatus.error],
   /// otherwise null.
-  final String errorMessage;
+  final String? errorMessage;
 
   FacebookLoginResult._(Map<String, dynamic> map)
       : status = _parseStatus(map['status']),
@@ -229,7 +229,7 @@ class FacebookLoginResult {
             : null,
         errorMessage = map['errorMessage'];
 
-  static FacebookLoginStatus _parseStatus(String status) {
+  static FacebookLoginStatus _parseStatus(String? status) {
     switch (status) {
       case 'loggedIn':
         return FacebookLoginStatus.loggedIn;
@@ -264,10 +264,10 @@ enum FacebookLoginStatus {
 class FacebookAccessToken {
   /// The access token returned by the Facebook login, which can be used to
   /// access Facebook APIs.
-  final String token;
+  final String? token;
 
   /// The id for the user that is associated with this access token.
-  final String userId;
+  final String? userId;
 
   /// The date when this access token expires.
   final DateTime expires;
@@ -277,14 +277,14 @@ class FacebookAccessToken {
   /// These are the permissions that were requested with last login, and which
   /// the user approved. If permissions have changed since the last login, this
   /// list might be outdated.
-  final List<String> permissions;
+  final List<String>? permissions;
 
   /// The list of declined permissions associated with this access token.
   ///
   /// These are the permissions that were requested, but the user didn't
   /// approve. Similarly to [permissions], this list might be outdated if these
   /// permissions have changed since the last login.
-  final List<String> declinedPermissions;
+  final List<String>? declinedPermissions;
 
   /// Is this access token expired or not?
   ///
