@@ -41,9 +41,9 @@ void main() {
     };
 
     final log = <MethodCall>[];
-    FacebookLogin sut;
+    late FacebookLogin sut;
 
-    void setMethodCallResponse(Map<String, dynamic> response) {
+    void setMethodCallResponse(Map<String, dynamic>? response) {
       channel.setMockMethodCallHandler((methodCall) {
         log.add(methodCall);
         return Future.value(response);
@@ -94,7 +94,7 @@ void main() {
       setMethodCallResponse(kLoggedInResponse);
 
       final result = await sut.logIn([]);
-      final map = result.accessToken.toMap();
+      final map = result.accessToken!.toMap();
 
       expect(
         map,
@@ -128,7 +128,7 @@ void main() {
       setMethodCallResponse(null);
 
       // Setting a null login behavior is not allowed.
-      expect(() => sut.loginBehavior = null, throwsAssertionError);
+      // expect(() => sut.loginBehavior = null, throwsAssertionError);
     });
 
     test('loginBehavior - nativeWithFallback is the default', () async {
@@ -187,7 +187,7 @@ void main() {
       ]);
 
       expect(result.status, FacebookLoginStatus.loggedIn);
-      expectAccessTokenParsedCorrectly(result.accessToken);
+      expectAccessTokenParsedCorrectly(result.accessToken!);
 
       expect(
         log,
@@ -278,7 +278,8 @@ void main() {
         () async {
       setMethodCallResponse(kAccessToken);
 
-      final accessToken = await sut.currentAccessToken;
+      final accessToken =
+          await (sut.currentAccessToken as Future<FacebookAccessToken>);
       expectAccessTokenParsedCorrectly(accessToken);
     });
 
@@ -288,7 +289,8 @@ void main() {
 
       Clock.dateTimeResolver = () => beforeExpiry;
 
-      final accessToken = await sut.currentAccessToken;
+      final accessToken =
+          await (sut.currentAccessToken as Future<FacebookAccessToken>);
       expect(accessToken.isValid(), isTrue);
     });
 
@@ -298,7 +300,8 @@ void main() {
 
       Clock.dateTimeResolver = () => afterExpiry;
 
-      final accessToken = await sut.currentAccessToken;
+      final accessToken =
+          await (sut.currentAccessToken as Future<FacebookAccessToken>);
       expect(accessToken.isValid(), isFalse);
     });
   });
