@@ -21,7 +21,7 @@ class _MyAppState extends State<MyApp> {
 
     switch (result.status) {
       case FacebookLoginStatus.loggedIn:
-        final FacebookAccessToken accessToken = result.accessToken;
+        final FacebookAccessToken accessToken = result.accessToken!;
         _showMessage('''
          Logged in!
          
